fix testoptions error
