+++ alchemy/src/test/bun.ts
@@ -4,8 +4,8 @@ import { afterAll, beforeAll, it } from "bun:test";
 import path from "node:path";
 import { alchemy } from "../alchemy.ts";
 import { Scope } from "../scope.ts";
-import type { StateStoreType } from "../state.ts";
 import { NoopTelemetryClient } from "../util/telemetry/index.ts";
+import type { TestOptions } from "./options.ts";
 
 /**
  * Extend the Alchemy interface to include test functionality
@@ -20,39 +20,6 @@ declare module "../alchemy.ts" {
  * Add test functionality to alchemy instance
  */
 alchemy.test = test;
-
-/**
- * Options for configuring test behavior
- */
-export interface TestOptions {
-  /**
-   * Whether to suppress logging output.
-   * @default false.
-   */
-  quiet?: boolean;
-
-  /**
-   * Password to use for test resources.
-   * @default "test-password".
-   */
-  password?: string;
-
-  /**
-   * Override the default state store for the test.
-   */
-  stateStore?: StateStoreType;
-
-  /**
-   * Prefix to use for the scope to isolate tests and environments.
-   */
-  prefix?: string;
-
-  /**
-   * Apply updates to resources even if there are no changes.
-   * @default false
-   */
-  force?: boolean;
-}
 
 /**
  * Test function type definition with overloads
